Split CyclopsHelmHUDManager patches.
Could do that as well while I was at it.
