+++ NitroxPatcher/NitroxPatcher.csproj
@@ -63,7 +63,9 @@
     <Compile Include="Patches\CrafterLogic_TryPickup_Patch.cs" />
     <Compile Include="Patches\Creature_ChooseBestAction_Patch.cs" />
     <Compile Include="Patches\CyclopsEngineChangeState_OnClick_Patch.cs" />
-    <Compile Include="Patches\CyclopsHelmHUDManager_Patch.cs" />
+    <Compile Include="Patches\CyclopsHelmHUDManager_Start_Patch.cs" />
+    <Compile Include="Patches\CyclopsHelmHUDManager_StopPiloting_Patch.cs" />
+    <Compile Include="Patches\CyclopsHelmHUDManager_Update_Patch.cs" />
     <Compile Include="Patches\CyclopsHornButton_OnPress_Patch.cs" />
     <Compile Include="Patches\CyclopsLightingPanel_ToggleFloodlights_Patch.cs" />
     <Compile Include="Patches\CyclopsLightingPanel_ToggleInternalLighting_Patch.cs" />
