+++ NitroxPatcher/Patches/CyclopsHelmHUDManager_StopPiloting_Patch.cs
@@ -1,0 +1,23 @@
+﻿using Harmony;
+using NitroxModel.Helper;
+using System;
+using System.Reflection;
+
+namespace NitroxPatcher.Patches
+{
+    public class CyclopsHelmHUDManager_StopPiloting_Patch : NitroxPatch
+    {
+        public static readonly Type TARGET_CLASS = typeof(CyclopsHelmHUDManager);
+        public static readonly MethodInfo TARGET_METHOD = TARGET_CLASS.GetMethod("StopPiloting", BindingFlags.Public | BindingFlags.Instance);
+
+        public static void Postfix(CyclopsHelmHUDManager __instance)
+        {
+            __instance.ReflectionSet("hudActive", true);
+        }
+
+        public override void Patch(HarmonyInstance harmony)
+        {
+            PatchPostfix(harmony, TARGET_METHOD);
+        }
+    }
+}
